function/requirements.txt
azure-functions
azure-storage-blob==12.17.0
azure-identity>=1.12.0